html/semantics/scripting-1/the-script-element/module/evaluation-order-2-import.html

<!DOCTYPE html>
<title>Testing evaluation order</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    window.log = [];

    window.addEventListener("error",
        ev => log.push("window-error", ev.error.message));
    window.addEventListener("onunhandledrejection", unreachable);

    const test_load = async_test("Test evaluation order of modules");

    window.addEventListener("load", ev => testDone());

    function unreachable() { log.push("unexpected"); }
    function testDone() {
      test_load.step(() => {
        assert_array_equals(log, [
            "step-2.2-1", "step-2.2-2",
            "window-error", "error",
            "microtask-2.2",
            "script-load",
          ]);
      });
      test_load.done();
    }
</script>

<script type="module" src="evaluation-order-2.1.mjs"
    onerror="unreachable()" onload="log.push('script-load')">
 </script>
